Proposal: COSE keys discovery endpoint
This is a late addition, but one that was discussed at IETF123 and seemed to receive positive feedback. It is strictly an addition, and does have a normative impact on any other endpoints in its current form.

We could go a little further, and define it as `.well-known/transparency-keys`, to separate it from the configuration. That would mean two well-known registration, but would avoid a request for a verifier that does not have the keys cached. What do we think?